cheerlights_discord/requirements.txt
paho-mqtt
discord-webhook